requirements.txt
aiogram==2.15
aiohttp==3.7.4.post0
aiosqlite==0.17.0
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
certifi==2021.5.30
chardet==4.0.0
idna==3.2
iso8601==0.1.16
more-itertools==8.10.0
multidict==5.1.0
pyaes==1.6.1
pyasn1==0.4.8
pypika-tortoise==0.1.1
pytz==2021.1
rsa==4.7.2
Telethon==1.23.0
tortoise-orm==0.17.7
typing-extensions==3.10.0.2
yarl==1.6.3
